Moved card spwawning to DisplayManager
