+++ Dixit/Assets/Scenes/MainScene.unity
@@ -452,7 +452,7 @@ PrefabInstance:
     - target: {fileID: 6398685927968754599, guid: f88226e001e7c6b478b8ff5f1177e5d8,
         type: 3}
       propertyPath: m_IsActive
-      value: 1
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: 6846551886779996817, guid: f88226e001e7c6b478b8ff5f1177e5d8,
         type: 3}
